Create excel table using jexcel library

This function is used to create excel like table

excelTable(data = NULL, columns = NULL, colHeaders = NULL,
  rowHeight = NULL, nestedHeaders = NULL, defaultColWidth = NULL,
  minDimensions = NULL, columnSorting = TRUE, columnDrag = FALSE,
  columnResize = TRUE, rowResize = FALSE, rowDrag = TRUE,
  editable = TRUE, allowInsertRow = TRUE, allowInsertColumn = TRUE,
  allowDeleteRow = TRUE, allowDeleteColumn = TRUE,
  allowRenameColumn = TRUE, allowComments = FALSE, wordWrap = FALSE,
  selectionCopy = TRUE, mergeCells = NULL, search = FALSE,
  pagination = NULL, fullscreen = FALSE, lazyLoading = FALSE,
  loadingSpin = FALSE, style = NULL)

Arguments

data
a data object, can either be dataframe or a matrix.

columns
a dataframe containing different column attributes. The row number of the dataframe specifies the column for which the attribute is to be specified and the header of the dataframe specifies the attribute.The following attributes are supported:

title: a string specifying title of the column or the column header.

width: a numerical value specifying the width of the column.

type: the type of the column. It can be one of text, numeric, hidden, dropdown, autocomplete, checkbox, radio, calendar, image or color.

source: a vector of options for column when the type is 'dropdwown'.

multiple: a boolean value indicating if the multiple options can be selected when the type is 'dropdown'. The default value is false.

colHeaders
a vector of specifying the column headers. If both 'colHeaders' and 'title' attribute in columns is specified, the latter will take precedence.

rowHeight
a dataframe or matrix specifying height of different rows. The first column is consists of numerical value the specifiesthe row number while the second column is also numerical value that specifies the height in pixels.

nestedHeaders
a list of dataframe having title and colspan as the attributes. The nested header in the same level should be in the same dataframe.

defaultColWidth
a numeric value specifying the default width of column when the width attribute of column is not specified. The default value is 50.

minDimensions
a vector that defines the minimum number of rows and columns in the tables irrespective of the size of data. The first parameter should be of number of columns followed by number of rows. If data is null then it will create an empty table with the given dimensions.

columnSorting
a boolean value indicating if column sorting should be enabled. When enabled double click on the table headers sorts the column. By default it is set to true.

columnDrag
a boolean value indicating if column dragging is enabled. By default it is set to false.

columnResize
a boolean value indicating if column resizing is enabled. By default it is set to true.

rowResize
a boolean value indicating if row resizing is enabled. By default it is set to false.

rowDrag
a boolean value indicating if rowDragging is enabled or not. By default it is set to true.

editable
a boolean value indicating if the table can be edited. By default it is set to true

allowInsertRow
a boolean value indicating if user is allowed to insert new rows. By default it is set to true.

allowInsertColumn
a boolean value indicating if user is allowed to insert a new column. By default it is set to true.

allowDeleteRow
a boolean value indicating if user is allowed to delete a row. By default it is set to true.

allowDeleteColumn
a boolean value indicating if user is allowed to delete a column. By default it is set to true.

allowRenameColumn
a boolean value indicating if user is allowed to rename the columns. By default it is set to true

allowComments
a boolean value indicating if commenting on cell should be enabled. By default it is set to false.

wordWrap
a boolean value indicating if words in the cells should wrap. By default it is set to false.

selectionCopy
a boolean value indicating if user is allowed to copy selected cells. By default it is set to true.

mergeCells
a list containing vector of colspan and rowspan respectively such that the tag of the list specifies the cell number.

search
a boolean value indicating if search should be enabled. By default it is set to false.

pagination
a numeric value indicating number of rows in a single page. If the data does not fit in a single page pagination is enabled.

fullscreen
a boolean value indicating if the table should be fullscreen. By default it is set to false.

lazyLoading
a boolean value indicating if lazy loading should be enabled. By default it is set to false.

loadingSpin
a boolean value indicating if loading spinner should be enabled. By default it is set to false.

style
a list to specify style for each cell. The tag should be the cell value and the value should be a vector of css styles to be applied.

Examples

library(excelR)

data = data.frame(Model = c('Mazda', 'Pegeout', 'Honda Fit', 'Honda CRV'),
                  Date=c('2006-01-01', '2005-01-01','2004-01-01', '2003-01-01' ),
                  Availability = c(TRUE, FALSE, TRUE, TRUE))

columns = data.frame(title=c('Model', 'Date', 'Availibility'),
                     width= c(300, 300, 300),
                     type=c('text', 'calendar', 'checkbox'))

excelTable(data=data, columns = columns)